+++ components/Flagging/AddReport.tsx
@@ -4,7 +4,7 @@
  */
 "use client";
 
-import { useState, type ChangeEvent, type FormEvent } from "react";
+import { useState, type ChangeEvent, type SubmitEvent } from "react";
 import GeneralButton from "@components/ReusableSmallComponents/buttons/GeneralButton";
 import { toast } from "react-toastify";
 import axios from "axios";
@@ -63,7 +63,7 @@ export default function AddReport({
         );
   };
 
-  const handleSubmitReport = async (e: FormEvent<HTMLFormElement>) => {
+  const handleSubmitReport = async (e: SubmitEvent) => {
     e.preventDefault();
     setLoading(true);